refactor(personalization): move theme models from interface to model
1. Move ThemeVieweModel class from PersonalizationInterface to ThemeModel
2. Move globalThemeViewModel, iconThemeViewModel, cursorThemeViewModel to PersonalizationModel
3. Move appearanceSwitchModel and currentAppearance from PersonalizationInterface to PersonalizationModel
4. Move initAppearanceSwitchModel() from PersonalizationInterface to PersonalizationWorker
5. Update QML bindings to use dccData.model.xxx instead of dccData.xxx
6. Handle TreeLand window theme type changes to update current appearance

Log: refactor personalization module, move theme view models and appearance state from interface to model layer

refactor(personalization): 将主题模型从接口层移至模型层

1. 将 ThemeVieweModel 类从 PersonalizationInterface 移至 ThemeModel
2. 将 globalThemeViewModel、iconThemeViewModel、cursorThemeViewModel 移至 PersonalizationModel
3. 将 appearanceSwitchModel 和 currentAppearance 从 PersonalizationInterface 移至 PersonalizationModel
4. 将 initAppearanceSwitchModel() 从 PersonalizationInterface 移至 PersonalizationWorker
5. 更新 QML 绑定路径，使用 dccData.model.xxx 替代 dccData.xxx
6. 处理 TreeLand 窗口主题类型变化以更新当前外观状态

Log: 重构个性化模块，将主题视图模型和外观状态从接口层移至模型层
PMS: BUG-346357
